src/AppLogistics.Controllers/Configuration/Clients/ClientsController.cs
using AppLogistics.Components.Mvc;
using AppLogistics.Objects;
using AppLogistics.Services;
using AppLogistics.Validators;
using Microsoft.AspNetCore.Mvc;

namespace AppLogistics.Controllers.Configuration
{
    [Area("Configuration")]
    public class ClientsController : ValidatedController<IClientValidator, IClientService>
    {
        public ClientsController(IClientValidator validator, IClientService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create([BindExcludeId] ClientCreateEditView client)
        {
            if (!Validator.CanCreate(client))
            {
                return View(client);
            }

            Service.Create(client);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(int id)
        {
            return NotEmptyView(Service.Get<ClientView>(id));
        }

        [HttpGet]
        public ActionResult Edit(int id)
        {
            return NotEmptyView(Service.Get<ClientCreateEditView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ClientCreateEditView client)
        {
            if (!Validator.CanEdit(client))
            {
                return View(client);
            }

            Service.Edit(client);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(int id)
        {
            return NotEmptyView(Service.Get<ClientView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(int id)
        {
            if (!Validator.CanDelete(id))
            {
                return RedirectToAction("Delete", new { id });
            }

            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
